TINY-8778: Updated tests
