+++ modules/tinymce/src/core/main/ts/textpatterns/utils/Utils.ts
@@ -3,6 +3,7 @@ import { Optional } from '@ephox/katamari';
 import DOMUtils from '../../api/dom/DOMUtils';
 import Editor from '../../api/Editor';
 import * as NodeType from '../../dom/NodeType';
+import { Pattern, PatternSet } from '../core/PatternTypes';
 
 const cleanEmptyNodes = (dom: DOMUtils, node: Node | null, isRoot: (e: Node) => boolean): void => {
   // Recursively walk up the tree while we have a parent and the node is empty. If the node is empty, then remove it.
@@ -36,8 +37,16 @@ const deleteRng = (dom: DOMUtils, rng: Range, isRoot: (e: Node) => boolean, clea
 const getParentBlock = (editor: Editor, rng: Range): Optional<Element> =>
   Optional.from(editor.dom.getParent(rng.startContainer, editor.dom.isBlock));
 
+const resolveFromDynamicPatterns = (patternSet: PatternSet, block: Element): Pattern[] => {
+  const blockText = block.textContent ?? '';
+  return patternSet.dynamicPatternsLookup({
+    text: blockText,
+    block
+  });
+};
 export {
   cleanEmptyNodes,
   deleteRng,
-  getParentBlock
+  getParentBlock,
+  resolveFromDynamicPatterns
 };
