+++ modules/tinymce/src/core/test/ts/browser/textpatterns/FindInlinePatternTest.ts
@@ -5,11 +5,11 @@ import { assert } from 'chai';
 
 import Editor from 'tinymce/core/api/Editor';
 import * as InlinePattern from 'tinymce/core/textpatterns/core/InlinePattern';
-import { InlinePattern as InlinePatternType, InlinePatternMatch, InlinePatternSet } from 'tinymce/core/textpatterns/core/PatternTypes';
+import { InlinePattern as InlinePatternType, InlinePatternMatch, InlinePatternSet, PatternSet } from 'tinymce/core/textpatterns/core/PatternTypes';
 import { PathRange } from 'tinymce/core/textpatterns/utils/PathRange';
 import ListsPlugin from 'tinymce/plugins/lists/Plugin';
 
-import { getPatternSetFor } from '../../module/test/TextPatternsUtils';
+import { findPatternsWithDynamicPatterns, getPatternSetFor } from '../../module/test/TextPatternsUtils';
 
 describe('browser.tinymce.textpatterns.FindInlinePatternTest', () => {
   interface ExpectedPatternMatch {
@@ -18,8 +18,7 @@ describe('browser.tinymce.textpatterns.FindInlinePatternTest', () => {
     readonly endRng: PathRange;
   }
 
-  const getInlinePattern = (editor: Editor, patternSet: InlinePatternSet, space: boolean = false, normalized: boolean = false) =>
-    InlinePattern.findPatterns(editor, patternSet, normalized, space);
+  const getInlinePattern = (editor: Editor, patternSet: PatternSet, space: boolean = false, normalized: boolean = false) => findPatternsWithDynamicPatterns(editor, patternSet, InlinePattern.findPatterns, space, normalized);
 
   const assertPatterns = (actualMatches: InlinePatternMatch[], expectedMatches: ExpectedPatternMatch[]) => {
     assert.lengthOf(actualMatches, expectedMatches.length, 'Pattern count does not match');
@@ -217,21 +216,21 @@ describe('browser.tinymce.textpatterns.FindInlinePatternTest', () => {
       it('TINY-8778: force only ** pattern and test return on not existing *** pattern', () => {
         const editor = hook.editor();
         setContentAndCursor(editor, '***x***', [ 0 ], 7);
-        const matches = getInlinePattern(editor, inlinePatternSet);
+        const matches = getInlinePattern(editor, inlinePatternSet as PatternSet);
         assertSimpleMatch(matches, '**', '**', [ 'bold' ], { start: [ 0, 1 ], end: [ 0, 3 ] }, { start: [ 0, 5 ], end: [ 0, 7 ] });
       });
 
       it('TINY-8778: force only ** pattern with leading content and test return on not existing *** pattern', () => {
         const editor = hook.editor();
         setContentAndCursor(editor, 'y ***x***', [ 0 ], 9);
-        const matches = getInlinePattern(editor, inlinePatternSet);
+        const matches = getInlinePattern(editor, inlinePatternSet as PatternSet);
         assertSimpleMatch(matches, '**', '**', [ 'bold' ], { start: [ 0, 3 ], end: [ 0, 5 ] }, { start: [ 0, 7 ], end: [ 0, 9 ] });
       });
 
       it('TINY-8778: force only ** pattern with trailing ** text and test return on not existing *** pattern', () => {
         const editor = hook.editor();
         setContentAndCursor(editor, 'y ***x*** **', [ 0 ], 9);
-        const matches = getInlinePattern(editor, inlinePatternSet);
+        const matches = getInlinePattern(editor, inlinePatternSet as PatternSet);
         assertSimpleMatch(matches, '**', '**', [ 'bold' ], { start: [ 0, 3 ], end: [ 0, 5 ] }, { start: [ 0, 7 ], end: [ 0, 9 ] });
       });
     });
